refactor(cli): Rename dispatchToSubCommand to dispatch-to-sub-command
